+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[FIX]** Display `Itinerary` as small if we don't provide date or subLabel in places
 - **[FIX]** Added background for transparent avatars in `TripCard`
 - **[FIX]** Update `Title` style in `TripCard`
 - **[FIX]** Update margins in `EmptyState` component